Added Empty implementation for Analyser Added Realty Subscription module Added Persistance library Added pom.xml files as far decide to use tomcat instead of play FIX: some values for krisha.kz, because they changed it on thier side

+++ sources/crawler/adverts-crawler-krisha/app/kz/aphion/adverts/crawler/krisha/mappers/KrishaAdvertMapper.java
@@ -66,10 +66,10 @@ public class KrishaAdvertMapper {
 					adverts.add(realty);
 				} else {
 					// NULL - это проблема, видимо что-то случилось
-					Logger.error("ATTENTION: Advert with id [" + (Long)rawAdvert.get("id") + "] wasn't processed!");
+					Logger.error("ATTENTION: Advert with id [" + rawAdvert.get("id") + "] wasn't processed!");
 				}
 			} catch (Exception e) {
-				Logger.error(e, "ATTENTION: Advert with id [" + (Long)rawAdvert.get("id") + "] wasn't processed because of errors!");
+				Logger.error(e, "ATTENTION: Advert with id [" + rawAdvert.get("id") + "] wasn't processed because of errors!");
 			}
 		}
